tests: comparaison script: run pycharm cleanup
